Work on login testing with minitest
SessionsControllerTest: It took some work to get the successful login test
working, because of the complexity of the login process and because the
tutorial uses rspec while I'm using minitest.  The main login test -
test_good_redirect - now succeeds.  Committing this work before cleaning up
slop left over from this work and deleting 'test_good_values', which fails
because it calls:
      assert_response :success, 'normal response expected'
but the response is a redirect rather than the :success (200, I think)
expected here.  The tutorial does:
controller.current_user.should == @user
controller.should be_signed_in

but there's no 'controller' method or attribute in this class (with
minitest) and I've not figured out how to find and check this state.  So
I've concluded that test_good_redirect is enough for now and
test_good_values can go.