Logout functionality
Logout/sign-out has been implemented; also, began work to make the links on
each page dynamic - e.g., if logged in, render a 'sign-out' link and if not
logged in, render a 'sign-in' link.  There's a problem with the signed_in?
query in the SessionsHelper module - when accessed in _header.html.erb it
always returns false, even after the user has logged in.  I'm making this
commit to save my work before grabbing the current version (from github) of
M. Hartl's code for the tutorial book - for the SessionsHelper class - to
see if this will fix the signed_in? problem.