css and users/index.html.erb improvements
Moved the css settings out of index.html.erb to custom.css via a "users"
class.  Fixed problem with different sizes of the (round) block for the
users list vs the (round) blocks above and below it for the pagination
information - the pagination block was smaller: The table used for the
users block had width set to 106%.  Changed to 100%.  Changed .container
width from 710px to 98%.  A couple other small improvements.