PeriodTypeSpec data-related changes
Changed db/schema.rb to create the period_type_specs table the rails way
instead of doing it elsewhere using SQL - The latter was done in order to
be able to define a foreign key from period_type_id to
period_types.period_type_id, but this scheme is proving too inconvenient -
not fitting into the rails way of doing things was causing problems, so
went back to schema.rb without SQL and no foreign key in the database.
(Actually, the period_types table is currently not used and probably won't
be - replace by the in-memory facility provided by PeriodTypeConstants.)

Added code in sample_data.rake to create some "PeriodTypeSpec"s associated
with some of the users, for testing in the browser on the dev database.

Period-type-spec tests and other changes related to testing
"PeriodTypeSpec"s.