Dropped period_types table and added PeriodTypeSpec.category.
database changes: Got rid of the period_types reference table - It's easier
to just use PeriodTypeConstants for the period-type id and name.  Got rid
of the setup to create the period_type_specs table via SQL (which was used
in order to create a foreign key in sqlite) and added a migration to
re-create the table the conventional rails way.  Added a category field to
the period_type_specs table; it is used to distinguish between use for
long-term/charting (i.e., start dates beginning up to many years or decades
ago) and use for short-term/analysis (i.e., relatively recent start dates).
Also added 'category' to the period_type_specs model and made other
category-related changes to it and added 2 related convenience methods to
the user model.

views and controllers: Changes dealing with the new category field,
including splitting the page (users/show.html.erb) listing the period-type
specs into 2 parts: specs for analysis at the top and specs for charting at
the bottom.  Also made category-related changes to the test files.