Adjust period-type-spec dates with respect to today
The previous behavior used the start_date and end_date attributes of
PeriodTypeSpec as is, which results in the situation where the start_date
(if not edited) gets further and further in the past as time progresses,
resulting in longer and longer time periods for requesting chart data and
analysis.  Assuming that most users would prefer the start-date to be,
essentially, now - n-days-ago (and the end-date usually nil - i.e., now -
so that the covered time period stays the same length), I changed the
behavior to use an "effective_start_date" and "effective_end_date" (for
cases where the end date is specified/not-nil) such that these dates roll
forward (based on today [now]) as the clock moves forward.  Changed the
appropriate views (show and per-type-spec/edit) to use this scheme by using
effective_start_date and effective_end_date (new queries of the
PeriodTypeSpec class) instead of start_date and end_date.