Login/out to/from the MAS server
Hooked in some logic for logging into and out of the MAS server via the
mas_client subsystem.  The PagesController 'home' action creates a
@mas_client object for the current user (which logs the user into the MAS
server if he is not already logged in) - if the user is signed in.
Information about the client (to be refined) is displayed via the view.
Added logic such that the sign-out process also logs the user out of the
MAS server (MasClient::logout) and deletes the associated MasSession from
the database.

Also, a bug fix: The intraday period-type names in PeriodTypeConstants were
wrong (e.g., 'one-minute' should be: '1-minute').  Fixed by removing the
constant definitions in this module and using the ones (via 'include')
TimePeriodTypeConstants (which resides in library/mas_client).