Got basic charting functionality working via google charts.
Added a form to the home/root page to select a symbol and period type.
Submit/view-chart posts to charts/index, which (via the ChartsController)
retrieves the tradable/stock data for the selected symbol/period-type from
the MAS server and displays it as a candlestock chart with google-charts.
Essentially the same form (select symbol/period-type) is also on the
charts/index page, allowing the user to choose a different symbol and/or
period-type to be charted.  Only charting of the o/h/l/c tradable data
so far - no moving averages or other indicators and no volume.

The javascript code to draw the chart with google-charts is in:
app/assets/javascripts/charts.js.coffee - It needs to be cleaned up a bit.
app/views/charts/index.html.erb invokes this functionality with:
<%= javascript_include_tag "charts" %>
The 'gon' gem was added and is used to conveniently package the tradable
data and retrieve it in charts.js.coffee.

Note that to allow charts.js.coffee to be loaded and used only for the
charts/index page, the following line had to be deleted from
app/assets/javascripts/application.js:

//= require_tree .

'post "charts/index"' was added to routes.rb to create the needed route to
charts/index.