Bug fix re. effective*date vs *date
Bug fix in PeriodTypeSpecsController: with the new functionality to
adjust the start_date and end_date based on the current
date and the updated_at date, the call to update the attributes in the
database:
@period_type_spec.update_attributes(params[:period_type_spec])
ended up using the ":effective_start[/end]_date" key, because that was used
in the associated form, which caused an update error because
effective_start[/end]_date are methods, not attributes - i.e., not in the DB.
The fix was to modify the date keys (e.g., effective_start_date ->
start_date) before calling update_attributes.

Also, I noticed that the effective start/end dates were not being sent to
the MAS server (i.e., the original start/end dates were used).  Fixed by
using a new class, PeriodTypeSpecAdapter, to wrap a PeriodTypeSpec instance
(via the adapter pattern) such that a call to 'start_date' invokes
effective_start_date (and similar for end_date).  Changed MasClientArgs to
transparently substitute an instance of this class for each
PeriodTypeSpec, which is then fed to the MasClient object created by
MasClientTools.