Time-delimited requests, charts improvement, bug fixes
Implemented use of the "time-delimited" version of tradable-data requests
so that the time-period/range for a particular request can be changed on
the fly without needing to change the MAS-server session configuration
(which currently, I believe, can only be done by logging out and
re-logging back into the MAS server).

Bug fix: PeriodTypeSpec.effective_start_date and effective_end_date cached
their return values for efficiency.  The cache became out of date (thus
displaying wrong values in the form) when an update action occurred that
caused the start_date, end_date attributes to be saved to the database.
Fixed by adding: 'after_save :clear_caches' to PeriodTypeSpec -
clear_caches sets the cache/instance-attributes to nil.

Bug fix: PeriodTypeSpec.adjusted_datetime, used by the
effective_start/end_date routines to an updated time value based on
'updated_at' and the current date, was - due to time-zone subtleties -
at certain times returning a value a day earlier than the correct date.
(E.g., when updated-at was today, the adjusted start_date value should be
the same as the original, but it was sometimes [when, I believe, CST/CDT-
today and UTC-today were a day apart] one day earlier.)

Charts view: Added selection-list of start/end date so that the user can
change the date range as well as symbol or period-type.  Added a couple
more helper_methods in the app. controller.  Made the start_year and
end_year, used by the form, configurable via configuration.rb.

Bug fix and accompanying UI design/implementation change: An exception was
triggered if the user chose, from the home page, to view a stock with a
period-type for which there was no associated period-type spec.  The user
wasn't allowed to get to the home page to select and view a stock if he
didn't yet have any period-type specs.  But once you have created one spec
(say, for period-type <y>) you can then get to the home page and try to
view a stock chart for period-type <z> - this would trigger the bug.  The
fix is to, in the controller code, if no spec is found for the selected
period-type, create one on the spot and save it to the database, and then
use it for the transaction.  The UI was then changed to allow the user to
unconditionally view the home page and from there select a symbol and
period-type and hit 'View chart'. If he has no period-type specs one will
now be created for the selected period type and the chart will be displayed
(with a new spec created).