tradable analysis, etc.
Implementation of tradable-event analysis: If the user clicks the "Run
analysis" button on the home page (after selecting some stocks and
analyzers), TradableAnalyzersController.index is called, which calls
mas_client.request_analysis for each stock with the chosen set of
analyzers; the browser is then directed to the tradable_analyzers/index
page, which (via app/views/tradable_analyzers/index.html.erb) produces a
table, presenting the resulting analysis events (with name, date).

Added a bit of logic to handle a failure to log in to the MAS server (e.g.,
if the server process is not running or has hung).

SessionsController.create: When the user logs in, a MAS client session is
started (by calling ApplicationController.mas_client), so that the client
session exists for the entire time the user is logged into
tradable-analysis-toolset web server.

Added an 'is_intraday' query to TradableAnalyzer so that the user can
choose between intraday and interday analysis (or she can choose both).

The 'analyzers' query of MasSession now returns a hash table, for which the
keys are the analyzer names and the values are the "TradableAnalyzer"s.