Fix bug: new user required mas_session
When a new user is created a mas_session record for the user needed to
be created and saved to the database (which requires a login to the MAS
server, since the mas_session requires the MAS session key for its
construction).  The bug was triggering this precondition from PagesController:
  pre :mas_session do current_user.mas_session != nil end
After the fix, the precondition is satisfied (in the newly added
"user user - home" test and with a manual test using a browser).