Bug fix: create new user when MAS server is down.
When a new user is created and the MAS server is hung or not running,
ApplicationController.mas_client fails with a nil result and sets
@error_msg.  UsersController.create was ignoring this nil result,
causing an exception and an unhelpful/useless error message to the
browser user.  Changed 'create' to check for nil result and fail
gracefully, with the error message in 'flash[:failure]'.  The result is
that for both MAS login failures and database (@user.save) failure,
the user again sees the "create login" page with an explanation for the
failure near the top of the page.