Added event-type column to analysis event table.
TradableAnalyzer.name includes the event-type name in parentheses, which
looks ugly - Fix: Add a 'description' query to TradableAnalyzer that
returns the name with the parenthesized part removed; and change
tradable_analyzers/index.html.erb to add a 3rd column to the table of
events: the event-type (name - i.e., the same part that was previously
in parens in the 1st column).  With some minor css adjustments, the
result definitely looks better than before.