Use new TradableTools to get stock name.
Made use of the new TradableTools class (in ./library/utility) to retrieve
the stock name based on the symbol.  This happens in ChartsController,
replacing the YahooFinance.quotes... call, which didn't work.  A global
instance of TradableTools is created in config/application.rb, using
'config.after_initialize do...' so that it is conveniently available
right after startup.  This instance is used in ChartsController.

Removed a few "marked" (with '!!!*') comments that are no longer needed -
e.g.: #!!!!!  attr_accessible :mas_session_key.

Also, changed bin/rails to not use spring, since spring is likely the
culprit for some unexplainable runtime errors observed recently.