Added Tradable/tradables model/table.
ChartsController, who is responsible for providing the stock/tradable
name in the 'index' method, if it is available, has been updated to
first query the database (the newly added tradables table) for the
name associated with the specified symbol; if the record exists, the
retrieved name is provided for the resulting 'charts/index' page
(stock chart, with name and symbol).  If no record is found in the
database, as before, the global $external_tradable_info
(TradableTools instance) is queried to get the name.  The resulting
name, of course, is then used for the 'chargs/index' page and also
saved to the database with the symbol.

Also: spring was removed from the Gemfile - It appeared to be causing
problems, perhaps due do the unorthodox layout of this app; and the
app/models classes were updated re. the new rails 5 scheme to inherit from
ApplicationRecord instead of ActiveRecord::Base.