Pre- & Post-condition-related bug fixes
PagesController: Added 'implies(signed_in?, @analyzers != nil)' to
postcondition of 'home' to protect line 21 of home.html.erb from
calling nil.count:
        <%= link_to "#{@analyzers.count} signal types",
and added code in 'analyzers_from_session' to enforce the
postcondition that result != nil and is a Hash (which, of course, helps
'home' enforce its new postcondition).
home.html.erb: TradableAnalyzersController#index has a precondition:
  current_user.mas_session.analyzers != nil
This was being violated when the current user does not yet have any
analyzers.  Added a guard in the form for tradable_analyzers_index_path
to hide the submit/"Run analysis" button if @analyzers is empty.