Use nonblocking/timeout facility (from mas_client)
Changes to make use of the new nonblocking/timeout behavior
with socket connections in library/mas_client:
Most of this logic went into MasClientTools::mas_client.  It is now
a recursive function - i.e., in these few lines (comments removed):
    result = MasClientNonblocking.new(@@mas_args)
    if result.communication_failed then
      if result.last_exception_type == MasTimeoutError then
        result = self.mas_client(session: session, user: user, next_port: true)
      ...
      end
    ...
if the MasClientNonblocking constructor fails due to a socket-
communication timeout error (e.g., took to long to connect or to
receive [read] the requested login response), mas_client is called
recursively with next_port true; and when next_port is true,
mas_client gets the next port in the configuration and tries to
connect (construct MasClientNonblocking) with it.  If the first
failure occurred because the MAS server at that port is
hosed/frozen/dead and the new port connects to a different MAS
server process, there's a fresh (and good) chance that this
connection/read attempt will succeed.  Needed to change the
ApplicationController, mas_client method to fit with these
changes to MasClientTools::mas_client, especially the new
next_port argument; and ChartsController.index needed to
check if 'client.communication_failed' and, if so, set the
flash error accordingly.  And, to support these changes in
MasClientTools, MasClientArgs needed some new features/functionality.

Also, added configuration of timeout-seconds value and MAS socket ports.