Try to fully employ the non-block/timeout facility
Attempt to consistently and completely make use of the new
nonblocking/timeout functionality so that most requests to the MAS
server can be recovered from - and tried again against a different
server process on a different port - if something goes wrong that
causes a time-out.  Changed almost all of the controller methods
that invoke a mas_client.request_<X> (send a request for <X> to the
MAS server and wait for a response) to check for a time-out or
related error and, with the help of a new utility method -
'do_mas_request' - in ApplicationController, in the case of such an
error, try to connect and make the request again on the "next port" -
iteratively until we've run out of (configured) ports.  Note: Some
more adjustment/tuning re. time-outs and errors is probably needed.

Changed MasClientTools::mas_client to no longer use the class
variable @@mas_args.  Instead, the only thing changeable that now
needs to be saved between sessions is
Rails.application.config.current_port_index (used to determine
which of the configured list of ports is the current port) and,
possibly, Rails.configuration.timeout_seconds - These are
components of the mas_args table.  (See MasClientArgs.)

Also, bug fix in app/views/tradable_analyzers/index.html.erb:  Put a
guard around @analysis_events[symbol] in case it is nil.  And a
small test fix in TradableAnalyzersTest.