analysis-events: models, save to DB, ...
Created new classes/models/tables needed for saving the results from
a request_analysis call to the MAS server in a well-structured format:
AnalysisEvent, AnalysisRun, tradableEventSet,
TradableProcessorParameterSetting, tradableProcessorRun.  Added logic
in the Analysis class to instantiate these classes with the correct
relationships, using the data returned from the MAS server.  Note
that the AnalysisEvent instances are actually instantiated by
TradableObjectFactory.new_event (which is called by the
MasCommunicationServices in the mas_client library).  These instances
are used (in Analysis) to create TradableEventSet objects, etc.
Added database logic in the Analysis class to save these objects to
the database at an appropriate time, within a transaction, with the
new (but obvious) design change to save the event data to the DB not
only for future review by the user, but also to allow notifications
about the events (via email, text, etc.) to be carried out by a
separate process (instead of the original idea to have Analysis trigger
notification via the Observable interface), thus freeing the "analysis"
process to focus only on MAS analysis requests and storing the resulting
responses.

Added some code/logic to AnalysisProfile and other related classes
(such as EventGenerationProfile) to fulfill some functionality
that was either needed by, or useful to, the new AnalysisEvent-related
code.

Also, made a couple changes/improvements to the locale/Chinese-translations
based on some feedback from a language partner in China.

Some changes/updates to the documentation, although it's probably
still a little out of date.

More test code, and some changes, for this new functionality (in
mas_client_test.rb).