NotificationAddress, AddressAssignment, ...
Implemented a many-to-many association between notification_addresses
(addresses used to send an email, text, or etc.) and users of such
addresses.  The association is polymorphic (i.e., a rails polymorphic
association) on the address-user side.  It is currently configured to
allow AnalysisProfile objects to have addresses (which will be used to
send notifications with analysis results when an analysis run is
triggered), as well as addresses for AnalysisSchedule objects.  The latter
association might be temporary - it was implemented to test that the
polymorphic association works, but it may not be needed in the final
version of the application.  For administration purposes, all
NotificationAddress objects are "owned" by a user, as well as being
"used" by any "address_users" that use it.

Other changes:
  - TradableObjectFactory: cleanup - removed obsolete code, etc.
  - Documentation updates.
  - Tests for the notification-address-related code.
  - Added a couple missing class-header comments/descriptions.