Exchange-close monitoring/notification
Implemented the following to-do item from the current
(2019-feb-services-iteration):
Design and implement a mechanism that monitors configured exchanges and, when
the closing-time for an exchange, or set of exchanges, passes determine which
symbols/tradables that belong to those exchanges are being tracked.  Then
notify the EOD data-retrieval service (via redis) to start looking for EOD
data for these symbols.  Two new classes created for implementation:
ExchangeScheduleMonitor and ExchangeClock, as well as two rake tasks
eod_exchange_monitoring_startup.rake and start_eod_exchange_monitoring.rake.
The implementation is mostly complete, except for the needed querying of
AnalysisProfiles to find the list of tracked tradables/symbols to send to the
EOD data-retrieval service; and it needs some cleanup and completion of a
couple other minor missing pieces of functionality, as well as tests/testing.

Other changes related to the above (exchange/EOD-monitoring):
  o Additional public methods for Exchange, used by ExchangeClock, with their
    corresponding implementation.
  o Additional interface/implementation, needed by the new code in Exchange,
    for MarketSchedule and MarketCloseDate.
  o New method: AnalysisProfile.in_use? - Will be used to find tracked
    tradables.
  o New methods: AnalysisSchedule.active?, User.active? - Called by
    AnalysisProfile.in_use?
  o TradableSymbol: Added:
    - association: belongs_to: :exchange
    - tracked?, track!, untrack!
  o db/stocks_with_exchange.csv (new list of stocks), updates to other csv
    files and to seeds.rb.
  o Some documentation updates.
  o Small database changes (db/migrate:
    20190210211756_add_exchanges_ref_to_tradable_symbols.rb,
    20190212222316_add_columns_to_exchanges.rb,
    20190213121745_add_columns_to_tradable_symbols.rb)
  o TatUtil.localtime
  o The new redis-based publish/subscribe-related modules/classess -
    ExchangeScheduleMonitor inherits from the new Publisher class.

Other changes, mostly related to the previous EOD-data retrieval/storage
management work:
  o DataConfig
  o Changed code that uses the redis publish/subscribe interface to use the
    new redis-based publish/subscribe modules/classes:
    - class EODRetrievalManager < PublisherSubscriber
  o Some related test changes/tests.