Discovery & cleanup of tracked tradables
The new class TradableTrackingManager is responsible for periodically
querying the database to find TradableSymbols that are being used that were
not being used before and, thus, marking them 'tracked'.  It also
periodically does a "clean-up" in order to change (formerly tracked)
TradableSymbols that are no longer used to NOT tracked (i.e., tracked =
false).  The goal is to allow the exchange monitor to do a quick, efficient
query on tradable_symbols rather than looking for all symbol_lists that
have an owner and then finding the TradableSymbols id.d by those lists,
etc.   In other words, the TradableTrackingManager does this so that the
ExchangeScheduleMonitor doesn't have to.  TradableTrackingManager also
coordinates-with/controls ExchangeScheduleMonitor so that it is not, e.g.,
in the middle of updating the database while the exchange monitor is
querying for tracked tradables.  The "tracking-management" functionality
also involved changes to:
  - AnalysisProfile: tracked_tradable_ids convenience function.
  - SymbolList: 'updated_since' scope that returns SymbolLists whose
    'updated_at' is > a specified time.
  - SymbolListAssignment: 'updated_since' added here, as well.
  - TradableSymbol/tradable_symbols: tracking_count -> tracked - We don't
    need to keep track of the count - just on (tracked) or off.
  - ForkedDatabaseExecution - some pruning/tuning, ....
  - RedisFacilities: Some new features used by TradableTrackingManager.
  - TatServicesFacilities: Several new features/utilities/constants/... used
    by TradableTrackingManager.
  - DataConfig: TRACKING_CLEANUP_INTERVAL = 61200
  - ExchangeScheduleMonitor: Changes to fit in with concepts introduced for
    TradableTrackingManager, obtained from module TatServicesFacilities -
    e.g., run_state, process_external_command, wait_for_resume_command, etc.
  - ExchangeClock: This is where the new, efficient query of tracked
    records from tradable_symbols, described above, is implemented, replacing
    the previous, more expensive and complicated scheme.
  - TradableSymbolTest: Change corresponding to changes in TradableSymbol.

Other changes: EODRetrievalManager: Changed to use RedisFacilities.