EOD retrieval: prep for thread -> process switch
Commit to save work before abandoning TimerTask-based (concurrent-ruby)
EOD data retrieval for a process-based solution.  The problem with the
threaded/TimerTask version was that it's very difficult to get hold of
exception information - i.e., I wasn't able to catch and examine any
exceptions within the TimerTask thread except for those that occurred at
the very top level (i.e., those that occur directly within the block
passed to the TimerTask constructor; any exceptions, e.g., within a
method called by that block, just disappeared).  Of course, I need this
exception info to debug a problem that appeared to be occurring during
the 'remove_from_set' or the 'add_set' call.  Using a separate process
to handle each case in which a set of exchanges close at the same time
will not take up much more resources than using threads, since
market-closes will not occur very often; and, of course, it should be
straightforward to obtain the needed debugging info.  Since there's some
intricate logic in this version, I decided to save it before switching in
case I might want to bring it back or use it for reference/example.

Also, reorganized/moved some of the manager classes - into one of two
categories: general or service_specific.

Classes affected by the thread-related work and/or the thread -> process
switch:
  - EODRetrievalManager: Moved the EODDataWrangler class into its own
    file.
  - Publisher, PublisherSubscriber, RedisFacilities, Subscriber,
    TatServicesFacilities: Changes related to the Redis handles (@redis
    and @redis_admin) made to deal with threading issues.  These changes
    might be reversed (might not be needed) after the switch.
  - FileTradableStorage: One or two small changes and a bunch of added,
    ugly debugging/puts code (which will be removed later).