EOD retrieval: Completion of threads -> processes
Completed implementation of the switch from multi-threaded to
multi-process for the EOD-data retrieval functionality - i.e., the switch
to processes described in the previous commit has been completed and
verified by testing.  Included in the implementation is handling of
network failure that can occur while attempting to retrieve EOD data from
the data provider, by detecting the error and, after waiting several
seconds, trying again.  Tested by simply unplugging the ethernet port of
the development computer and then, after several minutes, plugging it
back in.  Note that the retry logic was at first implemented by simply
looping within the "data-wrangler" child process until the retrieval
succeeded, but (due, it appears, to how the Net::HTTP package is
designed) when the network connection was restored connection attempts in
the code still failed - the restored connection was not detected.  This
was solved by changing the design to retry by (after exiting the initial
"data-wrangler" process) starting a new process for each retry attempt,
until a successful retrieval occurs.

Other changes:
  - seeds.rb: cleanup - removed some obsolete code.
  - TatServicesFacilities: A couple new messaging-utility methods.
  - Some general cleanup: removal of obsolete code, no-longer-needed
    puts/log.debug statements, etc.
  - DataWranglerHandler: A new class created (using the Concurrent::Async
    module) to manage the above-described data-retrieval-retry logic
    within asynchronous threads so that the main thread can continue to
    look for the next EOD-data-ready status and accompanying data
    retrieval (i.e., the next closing of an exchange).
  - ExchangeScheduleMonitor: Bug fix - Moved this line:
      eod_check_key = new_eod_check_key
    to inside of the method:
      def send_check_notification(symbols, closing_date_time)
    instead of the constructor ('initialize') - The 'eod_check_key' needs
    to be different each time a new "eod-check" event is published;
    otherwise, the lists of symbols for which EOD data is to be retrieved
    get mixed in with each other (i.e., with the same eod_check_key),
    breaking the data-retrieval model/logic set up in
    EODRetrievalManager.