Switch to queues for symbols/EOD-data processing
Changed to use - for the lists of symbols being passed between services
for processing (data retrieval/storage, analysis, etc.), instead of Redis
sets, queues implemented via Redis lists.  The reason for the change is
to attempt to make the passing of these symbol lists more robust so that,
for example, if a process updating two lists crashes during the update,
none of the data (symbols) stored in the lists is lost.  This is
accomplished via Redis' rpoplpush operation (used here to move from,
e.g., "queue A" to "queue B"), which is atomic.  These changes are seen,
so far in EODDataWrangler, which uses the new queueing logic and
RedisFacilities, which uses rpoplpush and other list operations to
implement the queue abstraction.  (Old/obsolete code in EODDataWrangler
still needs to be cleaned up.)

Other changes:
  - EODRetrievalManager: A couple small changes to fit in with the new
    queue-based scheme described above.
  - ExchangeScheduleMonitor: A small change for the new queue paradigm:
<      count = add_set(eod_check_key, symbols.map {|s| s.symbol},
<                     DEFAULT_EXPIRATION_SECONDS)
    becomes:
>      count = queue_messages(eod_check_key, symbols.map {|s| s.symbol},
>                      DEFAULT_EXPIRATION_SECONDS)
  - Upgraded rails from 5.0.7.1 to 5.0.7.2 to allow, as well, actionview
    upgrade to 5.0.7.2.  This is to address a security alert from github
    ( https://groups.google.com/forum/#!topic/rubyonrails-security/zRNVOUhKHrg )
  - ServiceTokens: Some new tokens; cleanup.
  - TatServicesFacilities: A few new EOD-data-related features, a couple
    of which are newly used by EODDataWrangler.  Etc.
  - FileTradableStorage, TradableDataRetriever: 'data_up_to_date_for' was
    enhanced to, if no data is yet available from the provider, the
    start-date used for 'symbol' (obtained from the new feature
    TradableDataRetriever.start_date_for) is checked and, if it is later
    than the 'date' argument, the result is true - In other words if the
    date of the day after the latest date in the stored file for 'symbol'
    (the start-date) is later than the date we want to get data for
    (i.e., essentially, end-date), we are already up to date.
  - RakeManager: Additional symbol-tags/task-names.
  - ServicesSupervisor: 'thread_report' for debugging; etc.