Replace direct use of Redis with MessageBroker
Replaced RedisFacilities with MessagingFacilities - i.e., changed from
Redis-dependent abstraction to a general messaging abstraction.  The
interface of MessagingFacilities is essentially the same as the old
RedisFacilities interface, but its implementation is delegated to the
'broker' and 'admin_broker' (for administrative-related messages)
attributes (which, via MessageBrokerConfiguration, end up being of type
RedisMessageBroker).

Changed the publish/subscribe classes and modules to define a
non-Redis-dependent interface, which (as with MessagingFacilities)
delegates its implementation to a "broker" (in this case,
'pubsub_broker').  pubsub_broker ends up being (via
MessageBrokerConfiguration) of type RedisPubSubBroker.  The
Redis-dependent code resides (as before) in
services/lib/architectural/redis/ and the new messaging-abstraction code
resides in the new location services/lib/architectural/messaging/.

RedisErrorLog has been moved to the 'redis' directory and direct
dependence on it has been removed.

RedisManager has been renamed to MessageBrokerManager and modified to
fulfil its new purpose of managing a generic message broker.

Any code not in the 'redis' cluster that used to depend directly on
Redis-specific code has either been removed or modified to depend on the
message-broker abstraction.  This includes the 'REDIS' constant in
ServiceTokens being changed to 'MESSAGE_BROKER'.  DataConfig now depends
on the new MessageBrokerConfiguration class to obtain new message-broker
(RedisMessageBroker) and pub/sub-broker (RedisPubSubBroker) instances.