Redis -> MessageBroker: options -> expire_secs
Completed this task from the iteration to-do list:
  * If possible, get rid of the 'options' arguments, which imply a
    dependence on Redis.
Replaced the 'options' argument (to the 'set_message' method) with
expire_secs - i.e., expiration/TTL time in seconds.  Changed the code
that calls set_message and any code that sets up what used to be the
options hash to fit with the new (simple numeric) semantics for
expiration seconds.