code reorganization - goal: hexagonal architecture
Reorganized/restructured the directory tree containing the code - mostly
the files residing in ./services.  One of the main goals with respect to
this reorganization is to put the domain model at the center of the
architecture, such that the domain model does not depend on the UI, the
database/ActiveRecord, or etc. - that is, the UI and infrastructure code
will depend on the domain model and not the other way around.  This
restructuring is not yet complete; what has been done so far includes:
  o Several files were moved out of ./services into a more appropriate
    (possibly new) location/directory.
  o New root directories (under $TATDIR):
    - domain
    - external
  o Other new directories:
    - config/non_rails
    - lib/util
    - lib/messaging
  o Configuration path adjustments (e.g., in config/application.rb) needed
    as a result of this restructuring.