intercommunication, TestExchangeScheduleMonitor...
TestExchangeScheduleMonitor, TestExchangeClock, TatServicesFacilities,
TatServicesConstants, EODCommunicationsFacilities: Test of the queuing
of EOD-check key and associated (tradable) symbols in response to
a 'next-close-time'-occurred event.  And related refactoring in
preparation for moving services-intercommunication logic out of
ExchangeScheduleMonitor to an *InterCommunications class (as was
done for EODRetrievalManager with respect to
EODRetrievalInterCommunications).

test_eod_exchange_monitoring.rake, eod-exchange-monitoring-test: Test
driver for TestExchangeScheduleMonitor - rake script and bash script
that runs it.

TestEODRetrievalManager: (as with EODRetrievalManager) changed direct
services-communications calls to use 'intercomm'
(EODRetrievalInterCommunications).

EODRetrievalInterCommunications: Refinements, improvements - including
adding of publishing (now inherits from Publisher) and getting rid of
the need for 'owner = self.owner' argument for most methods.

EODDataWrangler: Moved a few service-intercommuncation calls to the
'intercomm' that were missed in the previous version.

ExchangeScheduleMonitor: Small changes and notes/comments in preparation
for the move of service-communications logic into a *InterCommunications
class.