Services intercomm.: tradable-tracking, run-states
TAT::TradableTrackingManager, TradableTrackingManager: Moved
service-intercommunications logic into new class:
TradableTrackingInterCommunications.  Moved the initialization logic
that does not depend on the database from (the ActiveRecord-dependent)
TradableTrackingManager to TAT::TradableTrackingManager (the domain
module).

ServiceStateFacilities: New module allowing the implementation and use
of the service run-state-related logic (methods, constants, etc.) to
be moved from the service managers (ExchangeScheduleMonitor,
EODRetrievalManager, etc.) to the the services-communications subsystem,
since the logic for querying and control of a service's run-state can
be thought of as services-intercommunications logic.  This task, of
course, required changes to EODRetrievalManager, ExchangeScheduleMonitor,
StatusReporting, TradableTrackingManager.
(Note: EODEventManager [whose specs have not yet been implemented] has
not yet been modified to fit into this framework.)

TatServicesFacilities: Re the architectural changes described above,
a lot of code, constant definitions, etc. have been moved - much of it
into ServiceStateFacilities and TatServicesConstants.

ServiceManager: With the above changes, it now needs to include
ServiceStateFacilities for the 'is_alive?' method.

ServicesSupervisor: Temporarily removed start-up of event-triggering -
until EODEventManager can be brought up to date.  Moved status-reporting
to be started up first so that it is available for queries/reports ASAP.

TAT::ExchangeClock: method 'tracked_tradables': The :empty_if_0_exchanges
postcondition was being violated during a test run.  This appeared to be
caused by how the Contracts::DSL implementation was unpacking the block
arguments used in the postcondition, rather than a defect in the code.
"Fixed" the problem by removing the postcondition.